Use new env variables in xrt

+++ tools/xrt/src/targets/sim/ClockProducerThread.cpp
@@ -1,28 +1,35 @@
 #include "targets/sim/ClockProducerThread.hpp"
 
-ClockProducerThread::ClockProducerThread(DUT* dut, Sync& wp, Sync& wd, Sync& rd, Sync& other, Semaphore& exitTick ):
-        Thread("clockprodthread"),
-        pDUT(dut),
-        mWP(wp),
-        mWD(wd),
-        mRD(rd),
-        mOther(other),
-        mExitTick(exitTick){
-}
+#include <algorithm>
+#include <cstdint>
+
+ClockProducerThread::ClockProducerThread(
+    DUT* dut, Sync& wp, Sync& wd, Sync& rd, Sync& other, Semaphore& exitTick)
+    : Thread("clockprodthread"),
+      pDUT(dut),
+      mWP(wp),
+      mWD(wd),
+      mRD(rd),
+      mOther(other),
+      mExitTick(exitTick) {}
 
 void ClockProducerThread::run() {
-    std::cout << std::dec << "[" << pDUT->getTime() << "]" << " Started: " << getCurrentThreadName() << std::endl;
+    std::cout << std::dec << "[" << pDUT->getTime() << "]"
+              << " Started: " << getCurrentThreadName() << std::endl;
 
     std::map<unsigned int, unsigned int> map;
- 
-     while (mExitTick.available() == 0) {
+
+    while (mExitTick.available() == 0) {
         Thread::sleepm(1);
         uint64_t minRequestdTicks;
 
         minRequestdTicks = UINT64_MAX;
-        minRequestdTicks = std::min(minRequestdTicks, mWP.getRequestedTicks().available());
-        minRequestdTicks = std::min(minRequestdTicks, mWD.getRequestedTicks().available());
-        minRequestdTicks = std::min(minRequestdTicks, mRD.getRequestedTicks().available());
+        minRequestdTicks = std::min(
+            minRequestdTicks, static_cast<uint64_t>(mWP.getRequestedTicks().available()));
+        minRequestdTicks = std::min(
+            minRequestdTicks, static_cast<uint64_t>(mWD.getRequestedTicks().available()));
+        minRequestdTicks = std::min(
+            minRequestdTicks, static_cast<uint64_t>(mRD.getRequestedTicks().available()));
 
         while (minRequestdTicks-- > 0) {
             mWP.getRequestedTicks().acquire();
@@ -30,9 +37,10 @@ void ClockProducerThread::run() {
             mRD.getRequestedTicks().acquire();
             map = pDUT->get64Value("m00_axis_tdata");
 
-            for(auto& val : map)
-                if(val.first != 0xFFFF'FFFF && val.second != 0xFFFF'FFFF) {
-                    //std::cout << "WP WD RD Time: " << pDUT->getTime() << " values: " << val.first << ' ' << val.second << std::endl;
+            for (auto& val : map)
+                if (val.first != 0xFFFF'FFFF && val.second != 0xFFFF'FFFF) {
+                    // std::cout << "WP WD RD Time: " << pDUT->getTime() << " values: " <<
+                    // val.first << ' ' << val.second << std::endl;
                     masterValues.push_back(val.second);
                     masterValues.push_back(val.first);
                 }
@@ -44,17 +52,20 @@ void ClockProducerThread::run() {
         }
 
         minRequestdTicks = UINT64_MAX;
-        minRequestdTicks = std::min(minRequestdTicks, mWP.getRequestedTicks().available());
-        minRequestdTicks = std::min(minRequestdTicks, mRD.getRequestedTicks().available());
+        minRequestdTicks =
+            std::min(minRequestdTicks, mWP.getRequestedTicks().available());
+        minRequestdTicks =
+            std::min(minRequestdTicks, mRD.getRequestedTicks().available());
 
         while (minRequestdTicks-- > 0) {
             mWP.getRequestedTicks().acquire();
             mRD.getRequestedTicks().acquire();
             map = pDUT->get64Value("m00_axis_tdata");
 
-            for(auto& val : map)
-                if(val.first != 0xFFFF'FFFF && val.second != 0xFFFF'FFFF) {
-                    //std::cout << "WP RD Time: " << pDUT->getTime() << " values: " << val.first << ' ' << val.second << std::endl;
+            for (auto& val : map)
+                if (val.first != 0xFFFF'FFFF && val.second != 0xFFFF'FFFF) {
+                    // std::cout << "WP RD Time: " << pDUT->getTime() << " values: " <<
+                    // val.first << ' ' << val.second << std::endl;
 
                     masterValues.push_back(val.second);
                     masterValues.push_back(val.first);
@@ -65,17 +76,20 @@ void ClockProducerThread::run() {
         }
 
         minRequestdTicks = UINT64_MAX;
-        minRequestdTicks = std::min(minRequestdTicks, mWP.getRequestedTicks().available());
-        minRequestdTicks = std::min(minRequestdTicks, mWD.getRequestedTicks().available());
+        minRequestdTicks =
+            std::min(minRequestdTicks, mWP.getRequestedTicks().available());
+        minRequestdTicks =
+            std::min(minRequestdTicks, mWD.getRequestedTicks().available());
 
         while (minRequestdTicks-- > 0) {
             mWP.getRequestedTicks().acquire();
             mWD.getRequestedTicks().acquire();
             map = pDUT->get64Value("m00_axis_tdata");
 
-            for(auto& val : map)
-                if(val.first != 0xFFFF'FFFF && val.second != 0xFFFF'FFFF) {
-                    //std::cout << "WP WD Time: " << pDUT->getTime() << " values: " << val.first << ' ' << val.second << std::endl;
+            for (auto& val : map)
+                if (val.first != 0xFFFF'FFFF && val.second != 0xFFFF'FFFF) {
+                    // std::cout << "WP WD Time: " << pDUT->getTime() << " values: " <<
+                    // val.first << ' ' << val.second << std::endl;
 
                     masterValues.push_back(val.second);
                     masterValues.push_back(val.first);
@@ -89,9 +103,10 @@ void ClockProducerThread::run() {
             mWP.getRequestedTicks().acquire();
             map = pDUT->get64Value("m00_axis_tdata");
 
-            for(auto& val : map) {
+            for (auto& val : map) {
                 if (val.first != 0xFFFF'FFFF && val.second != 0xFFFF'FFFF) {
-                    //std::cout << "WP Time: " << pDUT->getTime() << " values: " << val.first << ' ' << val.second << std::endl;
+                    // std::cout << "WP Time: " << pDUT->getTime() << " values: " <<
+                    // val.first << ' ' << val.second << std::endl;
 
                     masterValues.push_back(val.second);
                     masterValues.push_back(val.first);
@@ -105,9 +120,10 @@ void ClockProducerThread::run() {
             mWD.getRequestedTicks().acquire();
             map = pDUT->get64Value("m00_axis_tdata");
 
-            for(auto& val : map)
-                if(val.first != 0xFFFF'FFFF && val.second != 0xFFFF'FFFF) {
-                    //std::cout << "WD Time: " << pDUT->getTime() << " values: " << val.first << ' ' << val.second << std::endl;
+            for (auto& val : map)
+                if (val.first != 0xFFFF'FFFF && val.second != 0xFFFF'FFFF) {
+                    // std::cout << "WD Time: " << pDUT->getTime() << " values: " <<
+                    // val.first << ' ' << val.second << std::endl;
 
                     masterValues.push_back(val.second);
                     masterValues.push_back(val.first);
@@ -120,9 +136,10 @@ void ClockProducerThread::run() {
             mRD.getRequestedTicks().acquire();
             map = pDUT->get64Value("m00_axis_tdata");
 
-            for(auto& val : map)
-                if(val.first != 0xFFFF'FFFF && val.second != 0xFFFF'FFFF) {
-                    //std::cout << "RD Time: " << pDUT->getTime() << " values: " << val.first << ' ' << val.second << std::endl;
+            for (auto& val : map)
+                if (val.first != 0xFFFF'FFFF && val.second != 0xFFFF'FFFF) {
+                    // std::cout << "RD Time: " << pDUT->getTime() << " values: " <<
+                    // val.first << ' ' << val.second << std::endl;
 
                     masterValues.push_back(val.second);
                     masterValues.push_back(val.first);
@@ -135,9 +152,10 @@ void ClockProducerThread::run() {
             mOther.getRequestedTicks().acquire();
             map = pDUT->get64Value("m00_axis_tdata");
 
-            for(auto& val : map)
-                if(val.first != 0xFFFF'FFFF && val.second != 0xFFFF'FFFF) {
-                    //std::cout << "Other Time: " << pDUT->getTime() << " values: " << val.first << ' ' << val.second << std::endl;
+            for (auto& val : map)
+                if (val.first != 0xFFFF'FFFF && val.second != 0xFFFF'FFFF) {
+                    // std::cout << "Other Time: " << pDUT->getTime() << " values: " <<
+                    // val.first << ' ' << val.second << std::endl;
 
                     masterValues.push_back(val.second);
                     masterValues.push_back(val.first);
@@ -146,5 +164,6 @@ void ClockProducerThread::run() {
             mOther.getGeneratedTicks().release();
         }
     }
-    std::cout << std::dec << "[" << pDUT->getTime() << "]" << " Ended: " << getCurrentThreadName() << std::endl;
+    std::cout << std::dec << "[" << pDUT->getTime() << "]"
+              << " Ended: " << getCurrentThreadName() << std::endl;
 }
